+++ src/resdk/resources/sample.py
@@ -30,10 +30,13 @@ class Sample(SampleUtilsMixin, BaseCollection):
     collection = DictResourceField(
         resource_class_name="Collection", access_type=FieldAccessType.WRITABLE
     )
-    data = QueryRelatedField("Data")
-    relations = QueryRelatedField("Relation")
-    annotations = QueryRelatedField("AnnotationValue")
-    predictions = QueryRelatedField("PredictionValue")
+    data = QueryRelatedField("Data", "entity")
+    relations = QueryRelatedField("Relation", "entity")
+    annotations = QueryRelatedField("AnnotationValue", "entity")
+    predictions = QueryRelatedField("PredictionValue", "entity")
+    variants = QueryRelatedField("Variant", "variant_calls__sample")
+    experiments = QueryRelatedField("VariantExperiment", "variant_calls__sample")
+    variant_calls = QueryRelatedField("VariantCall")
 
     def __init__(self, resolwe, **model_data):
         """Initialize attributes."""
@@ -50,6 +53,17 @@ class Sample(SampleUtilsMixin, BaseCollection):
         self._background = None
         self._is_background = None
         super().update()
+
+    @property
+    def latest_experiment(self):
+        """Get latest experiment."""
+        return self.experiments.filter(ordering="-timestamp", limit=1)[0]
+
+    def variants_by_experiment(self, experiment):
+        """Get variants for sample detected by the given experiment."""
+        return self.resolwe.variant.filter(
+            variant_calls__sample=self.id, variant_calls__experiment=experiment.id
+        )
 
     @property
     def background(self):